app01/requirements.txt
Flask==3.1.3
opentelemetry-distro==0.61b0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-instrumentation-flask==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
SQLAlchemy==2.0.43
PyMySQL==1.1.2
Faker
requests